Dynamic terrain cubes + Threadpools in worldgrid
Vertex generation in terrains now based on how many
cells we want to present. The parameters are changed according to this,
to include this cube size.
This enables to make distance based rendering.

Thread are now started in groups of 4. This way they
dont put suddenly too much work on the CPU.

Functional terrain has been taken out completely.